Before collapsed nodes

+++ libmdtrie/src/defs.h
@@ -14,7 +14,7 @@ typedef uint64_t level_t;
 typedef uint64_t symbol_t;
 typedef uint8_t dimension_t;
 typedef uint64_t point_t;
-typedef uint8_t representation_t;
+typedef uint16_t representation_t;
 // typedef std::vector<point_t> coordinates_t;
 typedef point_t * coordinates_t;
 typedef std::vector<uint64_t> density_array;
@@ -22,6 +22,8 @@ typedef std::vector<uint64_t> density_array;
 extern uint64_t dfuds_size;
 const preorder_t null_node = -1;
 extern uint64_t get_bit_count;
+// typedef unsigned long long int TimeStamp;
+// extern TimeStamp backtrace_time;
 
 template<dimension_t DIMENSION>
 class tree_block;
